Fix potential reporting of misleading errors
As part of its last responsibility, the dump_core() function sends a
signal to the running process and then returns. So far, the caller has
taken the mere fact that the call returned as a sign that something was
amiss -- after all, the process was still running. However signals are
delivered in an asynchronous manner and so the call may very well
return.
With this change we make sure to treat a successful return from the
dump_core() function as a success, mitigating potential error message
about failures to kill the process.